requirements.txt
addict==2.4.0
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
datasets==2.10.1
dill==0.3.6
filelock==3.9.0
fonttools==4.39.0
frozenlist==1.3.3
fsspec==2023.3.0
huggingface-hub==0.13.1
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
kaggle==1.5.13
kiwisolver==1.4.4
Markdown==3.4.1
markdown-it-py==2.2.0
matplotlib==3.7.1
mdurl==0.1.2
mmcls==0.25.0
mmcv-full==1.7.1
mmsegmentation==0.30.0
model-index==0.1.11
multidict==6.0.4
multiprocess==0.70.14
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
opencv-python==4.7.0.72
openmim==0.3.6
ordered-set==4.1.0
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
prettytable==3.6.0
pyarrow==11.0.0
Pygments==2.14.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
responses==0.18.0
rich==13.3.2
safetensors==0.3.0
scipy==1.10.1
six==1.16.0
tabulate==0.9.0
text-unidecode==1.3
timm==0.8.15.dev0
torch==1.13.1
torchvision==0.14.1
tqdm==4.65.0
typing-extensions==4.5.0
urllib3==1.26.14
wcwidth==0.2.6
xxhash==3.2.0
yapf==0.32.0
yarl==1.8.2
zipp==3.15.0
